[Decomp][Player] Cohesion split: extract spell system to PlayerSpell.cpp
